:recycle: 참여자 컨트롤 부분 코드 리팩토링
