+++ client/src/components/room/host/ParticipantController.tsx
@@ -1,48 +1,7 @@
 import { useSelector } from 'react-redux';
 import { RootState } from '@src/store';
-import {
-  Wrapper,
-  RowWrapper,
-  ProfileDiv,
-  ControlDiv,
-} from '@components/room/host/ParticipantController.style';
-
-import { HumanIcon, VideoIcon, MicIcon } from '@components/icons';
-import SettingToggle from '@components/setting/SettingToggle';
-
-const Participant = ({ sid, user, userDevices, turnOffOtherVideo, turnOffOtherAudio }) => {
-  const targetNick = user?.nickname;
-  const targetImg = user?.imgUrl;
-  const isVideoOn = userDevices?.isVideoOn;
-  const isAudioOn = userDevices?.isAudioOn;
-
-  const turnOffVideo = () => {
-    if (!isVideoOn) return;
-    turnOffOtherVideo({ sid, isVideoOn: false });
-  };
-
-  const turnOffAudio = () => {
-    if (!isAudioOn) return;
-    turnOffOtherAudio({ sid, isAudioOn: false });
-  };
-
-  return (
-    <RowWrapper>
-      <ProfileDiv>
-        {targetImg ? <img src={targetImg} alt="other_user_image" /> : <HumanIcon />}
-        <span>{targetNick || 'judangs'}</span>
-      </ProfileDiv>
-      <ControlDiv>
-        <SettingToggle
-          Icon={VideoIcon}
-          isDeviceOn={isVideoOn}
-          setIsDeviceOn={() => turnOffVideo()}
-        />
-        <SettingToggle Icon={MicIcon} isDeviceOn={isAudioOn} setIsDeviceOn={() => turnOffAudio()} />
-      </ControlDiv>
-    </RowWrapper>
-  );
-};
+import { Wrapper } from '@components/room/host/ParticipantController.style';
+import Participant from '@components/room/host/Participant';
 
 const ParticipantController = ({ turnOffOtherVideo, turnOffOtherAudio }) => {
   const users = useSelector((state: RootState) => state.room.users);
